Move @param.query tests to a standalone file
